backend/requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.9.1
aiosignal==1.4.0
altgraph @ file:///AppleInternal/Library/BuildRoots/4~CAP1ugDqYZ2ZVF_54thwSWnK-8L4LO5_Zcx-VcI/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
async-timeout==4.0.3
asyncio==3.4.3
attrs==25.3.0
beautifulsoup4==4.12.2
bitvector-for-humans==0.14.1
busylight-for-humans==0.28.0
cachetools==6.2.1
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.2.1
croniter==6.0.0
cryptography==46.0.3
cycler==0.12.1
deprecation==2.1.0
distro==1.9.0
duckduckgo_search==8.1.1
exceptiongroup==1.3.0
faiss-cpu==1.10.0
fastapi==0.121.3
filelock==3.17.0
fonttools==4.51.0
frozenlist==1.7.0
fsspec==2025.2.0
future @ file:///AppleInternal/Library/BuildRoots/4~CAP1ugDqYZ2ZVF_54thwSWnK-8L4LO5_Zcx-VcI/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hidapi==0.14.0
hpack==4.1.0
html2text==2025.4.15
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.4.0
Jinja2==3.1.5
jiter==0.11.1
joblib==1.4.2
json5==0.9.14
kiwisolver==1.4.5
loguru==0.7.2
lxml==6.0.2
macholib @ file:///AppleInternal/Library/BuildRoots/4~CAP1ugDqYZ2ZVF_54thwSWnK-8L4LO5_Zcx-VcI/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
networkx==3.2.1
numpy==1.26.4
openai==2.5.0
packaging==24.0
pillow==10.3.0
postgrest==2.24.0
primp==0.15.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.1.2
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
realtime==2.24.0
redis==7.0.1
regex==2024.11.6
requests==2.31.0
rich==13.7.0
rq==2.6.0
rsa==4.9.1
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==3.4.1
six @ file:///AppleInternal/Library/BuildRoots/4~CAP1ugDqYZ2ZVF_54thwSWnK-8L4LO5_Zcx-VcI/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.1
soupsieve==2.8
starlette==0.49.3
storage3==2.24.0
StrEnum==0.4.15
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
sympy==1.13.1
threadpoolctl==3.5.0
tk==0.1.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.3
twilio==9.8.6
typer==0.9.4
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.38.0
webcolors==1.13
websockets==15.0.1
yarl==1.20.1
zipp==3.23.0
